Split downloader into checksum and boundary-test sibling modules
Resolve the module-size finding on
`installer/src/dependency_binaries/install/downloader.rs` (was 965 lines)
with a behaviour-preserving Rust refactor. Every source file is now at or
below 400 lines.

- New `checksum.rs` (245 lines): moves `fetch_expected_checksum`,
  `parse_checksum_token`, `compute_sha256`, and `verify_archive_checksum`
  plus their tests and the `parsed` checksum-state constant. Only
  `fetch_expected_checksum` and `verify_archive_checksum` are exposed, as
  `pub(super)`. It reuses the shared `CATEGORY_CHECKSUM` category and
  `map_ureq_error`, both now `pub(super)` in `downloader`.
- New `downloader_boundary_tests.rs` (370 lines, `#[cfg(test)]`): moves the
  loopback HTTP test server and the end-to-end `download_from_urls`
  boundary tests. `download_from_urls` is now `pub(super)` so the sibling
  test module can drive it; the public API is unchanged.
- `downloader.rs` (396 lines): retains URL/agent setup, destination
  capability acquisition, archive transfer, workflow orchestration, and the
  capability/error-mapping unit tests.
- `install/mod.rs` declares `mod checksum;` and
  `#[cfg(test)] mod downloader_boundary_tests;`.

Behaviour is preserved exactly: lower-case normalization, malformed and
empty checksum handling, error variants and fields, bounded 8192-byte
buffering, and the structured tracing categories and states are all
carried over verbatim. No crate-public API is broadened; only narrow
`pub(super)` seams were added.
